added correct URI pathing:
/House
/House/{id}
/House/{id}/User
/House/{id}/User/{id}
/House/{id}/Room
/House/{id}/Room/{id}
/House/{id}/Room/{id}/Device
/House/{id}/Room/{id}/Device/{id}
/House/{id}/Room/{id}/Environment

abstracted baseclasses from the object-model
