+++ src/main/java/smartHouse/objectClasses/Device.java
@@ -1,24 +1,21 @@
-package smartHouse.objectModel;
+package smartHouse.objectClasses;
 
 /**
  * Created by Patrik Glendell on 02/10/15.
  */
 
-import smartHouse.resourceModel.DeviceResource;
-
-import javax.ws.rs.core.Response;
 import java.util.Date;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class Device implements DeviceResource {
+public class Device {
     private int id;
     private String name;
     private Boolean status;
     private ConcurrentHashMap<Date,Double> Data;
 
     public Device() {
-
     }
+
     protected int getId() {
         return id;
     }
@@ -49,30 +46,5 @@ public class Device implements DeviceResource {
 
     protected void setData(ConcurrentHashMap<Date,Double> data) {
         Data = data;
-    }
-
-    @Override
-    public Response createDevice() {
-        return null;
-    }
-
-    @Override
-    public Response updateDevice(int id) {
-        return null;
-    }
-
-    @Override
-    public Device getDevice(int id) {
-        return null;
-    }
-
-    @Override
-    public ConcurrentHashMap<Integer, Device> getAllDevices() {
-        return null;
-    }
-
-    @Override
-    public Response deleteDevice(int id) {
-        return null;
     }
 }
