+++ src/main/java/smartHouse/objectClasses/Environment.java
@@ -1,8 +1,5 @@
-package smartHouse.objectModel;
+package smartHouse.objectClasses;
 
-import smartHouse.resourceModel.EnvironmentResource;
-
-import javax.ws.rs.core.Response;
 import java.util.Date;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -10,7 +7,7 @@ import java.util.concurrent.ConcurrentHashMap;
  * Created by Patrik Glendell on 02/10/15.
  */
 
-public class Environment implements EnvironmentResource {
+public class Environment {
     private int id;
     private ConcurrentHashMap<Date, Double> energyConsumption;
     private ConcurrentHashMap<Date, Double> waterConsumption;
@@ -33,25 +30,5 @@ public class Environment implements EnvironmentResource {
 
     protected void setWaterConsumption(ConcurrentHashMap<Date, Double> waterConsumption) {
         this.waterConsumption = waterConsumption;
-    }
-
-    @Override
-    public Response createEnvironment() {
-        return null;
-    }
-
-    @Override
-    public Response updateEnvironment() {
-        return null;
-    }
-
-    @Override
-    public Environment getEnvironment() {
-        return null;
-    }
-
-    @Override
-    public Response deleteEnvironment() {
-        return null;
     }
 }
