+++ src/main/java/smartHouse/resourceInterfaces/DeviceResourceInterface.java
@@ -1,38 +1,36 @@
-package smartHouse.resourceModel;
+package smartHouse.resourceInterfaces;
 
-import smartHouse.objectModel.User;
+import smartHouse.objectClasses.Device;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.concurrent.ConcurrentHashMap;
 
-
 /**
  * Created by Patrik Glendell on 02/10/15.
  */
-
-public interface UserResource {
+public interface DeviceResourceInterface {
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    Response createUser();
+    Response createDevice();
 
     @PUT
     @Path("{id}")
     @Consumes(MediaType.APPLICATION_JSON)
-    Response updateUser(@PathParam("id") int id);
+    Response updateDevice(@PathParam("id") int id);
 
     @GET
     @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    User getUser(@PathParam("id") int id);
+    Device getDevice(@PathParam("id") int id);
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    ConcurrentHashMap<Integer,User> getAllUsers();
+    ConcurrentHashMap<Integer,Device> getAllDevices();
 
     @DELETE
     @Path("{id}")
-    Response deleteUser(@PathParam("id") int id);
+    Response deleteDevice(@PathParam("id") int id);
 }
